requirements.txt
crewai[tools]
crewai-tools
streamlit
requests
python-dotenv
litellm
